refactor: icon 이름 수정
