+++ frontend/src/components/BreedSearch.tsx
@@ -1,6 +1,6 @@
 import React, { ChangeEvent, Dispatch, SetStateAction, useState } from 'react';
 import Topbar from '@/components/common/Topbar';
-import TopBack from '@/assets/icons/ic-top-back.svg';
+import TopBack from '@/assets/icons/ic-arrow-left.svg';
 import Search from '@/assets/icons/ic-search.svg';
 import { DogRegInfo } from '@/pages/Join';
 import { useBreed } from '@/hooks/useBreed';
