+++ frontend/src/components/home/DogCard.tsx
@@ -1,9 +1,9 @@
 import { DogStatistic } from '@/models/dog';
 import { walkPercentFormat } from '@/utils/format';
-import Ic from '@/assets/icons/ic-arrow.svg';
+import Ic from '@/assets/icons/ic-arrow-right.svg';
 import Avatar from '@/components/common/Avatar';
 import DayIcon from '@/components/home/DayIcon';
-import Walk from '@/assets/icons/walk/walk.svg';
+import Walk from '@/assets/icons/ic-walk.svg';
 import Progressbar from '@/components/home/Progressbar';
 
 interface DogCardProps {
