+++ frontend/src/components/walk/WalkHeader.tsx
@@ -1,7 +1,7 @@
 import Topbar from '@/components/common/Topbar';
-import Ic from '@/assets/icons/ic-arrow.svg';
+import Ic from '@/assets/icons/ic-arrow-right.svg';
 import { useNavigate } from 'react-router-dom';
-import Notification from '@/assets/icons/notification.svg';
+import Notification from '@/assets/icons/ic-notification.svg';
 import { Divider } from '@/components/common/Divider';
 export default function WalkHeader() {
     const navigate = useNavigate();
