+++ frontend/src/pages/Join/index.tsx
@@ -2,12 +2,12 @@ import Topbar from '@/components/common/Topbar';
 import { getStorage } from '@/utils/storage';
 import React, { useMemo, useRef, useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
-import TopBack from '@/assets/icons/ic-top-back.svg';
+import TopBack from '@/assets/icons/ic-arrow-left.svg';
 import Agreements from '@/pages/Join/Agreements';
 import { Button } from '@/components/common/Button';
 import { Divider } from '@/components/common/Divider';
 import PetOwner from '@/pages/Join/PetOwner';
-import Cancel from '@/assets/icons/ic-top-cancel.svg';
+import Cancel from '@/assets/icons/ic-cancel.svg';
 import { storageKeys } from '@/constants';
 import { useAuth } from '@/hooks/useAuth';
 import 'react-image-crop/dist/ReactCrop.css';
