+++ frontend/src/pages/MyPage/DogDetail.tsx
@@ -2,8 +2,8 @@ import { ResponseDogs, ResponseRecentMonthStatistics } from '@/api/dog';
 import { getUploadUrl } from '@/api/upload';
 import EditPhoto from '@/assets/buttons/btn-edit-photo.svg';
 import CrownIcon from '@/assets/icons/ic-crown.svg';
-import Back from '@/assets/icons/ic-top-back.svg';
-import Edit from '@/assets/icons/ic-top-edit.svg';
+import Back from '@/assets/icons/ic-arrow-left.svg';
+import Edit from '@/assets/icons/ic-edit.svg';
 import Distance from '@/classes/Distance';
 import BreedSearch from '@/components/BreedSearch';
 import DeleteDogModal from '@/components/DeleteDogModal';
